+++ src-inst/flow_storm/api.clj
@@ -4,16 +4,15 @@
   (:require [flow-storm.tracer :as tracer]
             [flow-storm.runtime.taps :as rt-taps]
             [flow-storm.utils :refer [log] :as utils]
-            [flow-storm.instrument.forms :as inst-forms]
-            [flow-storm.instrument.namespaces :as inst-ns]
-            [flow-storm.instrument.runtime :as inst-rt]
+            [hansel.api :as hansel]
+            [hansel.instrument.runtime :refer [*runtime-ctx*]]
+            [hansel.instrument.utils :as inst-utils]
             [flow-storm.runtime.debuggers-api :as dbg-api]
             [flow-storm.runtime.events :as rt-events]
             [flow-storm.runtime.values :as rt-values]
             [flow-storm.json-serializer :as serializer]
             [flow-storm.remote-websocket-client :as remote-websocket-client]
             [flow-storm.runtime.indexes.api :as indexes-api]
-            [clojure.repl :as clj.repl]
             [clojure.string :as str]
             [clojure.stacktrace :as stacktrace]))
 
@@ -105,84 +104,144 @@
   (rt-taps/setup-tap!)
   (println "Remote Clojure runtime initialized"))
 
-(defn instrument-var
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Clojure instrumentation ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn instrument-var-clj
 
   "Instruments any var.
 
-  (instrument-var var-symb opts)
+    Lets say you are interested in debugging clojure.core/interpose you can do :
 
-  Lets say you are interested in debugging clojure.core/interpose you can do :
+  (instrument-var-clj clojure.core/interpose)
 
-  (instrument-var clojure.core/interpose {})
+  Then you can call :
 
-  #rtrace (interpose :a [1 2 3])
+  (interpose :a [1 2 3])
+
+  and it will show up in the debugger.
 
   Be careful instrumenting clojure.core functions or any functions that are being used
   by repl system code since can be called constantly and generate a lot of noise.
 
-  Use `uninstrument-var` to remove instrumentation.
+  Use `uninstrument-var-clj` to remove instrumentation.
 
   `opts` is a map that support :flow-id and :disable
-  See `instrument-forms-for-namespaces` for :disable"
+  See `instrument-namespaces-clj` for :disable"
 
-  ([var-symb] (instrument-var var-symb {}))
+  ([var-symb] (instrument-var-clj var-symb {}))
   ([var-symb config]
 
-   (let [form (some->> (clj.repl/source-fn var-symb)
-                       (read-string {:read-cond :allow}))
-         form-ns (find-ns (symbol (namespace var-symb)))]
-     (if form
+   (hansel/instrument-var-clj var-symb (merge (tracer/hansel-config config)
+                                              config))
+   (rt-events/publish-event! (rt-events/make-var-instrumented-event (name var-symb)
+                                                                    (namespace var-symb)))))
 
-       (binding [*ns* form-ns]
-         (inst-ns/instrument-and-eval-form form-ns form config)
-         (log (format "Instrumented %s" var-symb))
-         (rt-events/publish-event! (rt-events/make-var-instrumented-event (name var-symb)
-                                                                          (namespace var-symb))))
-
-       (log (format "Couldn't find source for %s" var-symb))))))
-
-(defn uninstrument-var
+(defn uninstrument-var-clj
 
   "Remove instrumentation given a var symbol.
 
-  (uninstrument-var var-symb)"
+  (uninstrument-var-clj var-symb)"
 
   [var-symb]
-  (let [ns-name (namespace var-symb)]
-    (binding [*ns* (find-ns (symbol ns-name))]
-      (let [form (some->> (clj.repl/source-fn var-symb)
-                          (read-string {:read-cond :allow}))
-            expanded-form (inst-forms/macroexpand-all macroexpand-1
-                                                      (fn [symb] (symbol (resolve symb)))
-                                                      form
-                                                      ::original-form)]
-        (if form
 
-          (if (inst-forms/expanded-def-form? expanded-form)
-            (let [[v vval] (inst-ns/expanded-defn-parse ns-name expanded-form)]
-              (alter-var-root v (fn [_] (eval vval)))
-              (log (format "Uninstrumented %s" v))
-              (rt-events/publish-event! (rt-events/make-var-uninstrumented-event (name var-symb)
-                                                                                 (namespace var-symb))))
+  (hansel/uninstrument-var-clj var-symb)
+  (rt-events/publish-event! (rt-events/make-var-uninstrumented-event (name var-symb)
+                                                                     (namespace var-symb))))
 
-            (log (format "Don't know howto untrace %s" (pr-str expanded-form))))
+(defn instrument-namespaces-clj
 
-          (log (format "Couldn't find source for %s" var-symb)))))))
+  "Instrument all forms, in all namespaces that matches `prefixes`.
+
+    `prefixes` is a set of ns prefixes like #{\"cljs.compiler\" \"cljs.analyzer\"}
+
+  `opts` is a map containing :
+       - :excluding-ns  a set of strings with namespaces that should be excluded
+       - :disable a set containing any of #{:expr :binding :anonymous-fn}
+                  useful for disabling unnecesary traces in code that generate too many traces
+       - :verbose? when true show more logging
+  "
+
+  ([prefixes] (instrument-namespaces-clj prefixes {}))
+  ([prefixes opts] (dbg-api/instrument-namespaces :clj prefixes opts)))
+
+(defn uninstrument-namespaces-clj
+
+  "Undo instrumentation made by `flow-storm.api/instrument-namespaces-clj`"
+
+  [prefixes]
+  (dbg-api/uninstrument-namespaces :clj prefixes {}))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; ClojureScript instrumentation ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn instrument-var-cljs
+
+  "Like `flow-storm.api/instrument-var-clj` but for using it from the shadow Clojure repl.
+
+  Arguments are the same as the Clojure version but `config` also accepts a `:build-id`"
+
+  ([var-symb] (instrument-var-clj var-symb {}))
+  ([var-symb config]
+
+   (hansel/instrument-var-shadow-cljs var-symb (merge (tracer/hansel-config config)
+                                                      config))
+   (dbg-api/publish-event! :cljs
+                           (rt-events/make-var-instrumented-event (name var-symb)
+                                                                  (namespace var-symb))
+                           config)))
+
+(defn uninstrument-var-cljs
+
+  "Like `flow-storm.api/uninstrument-var-clj` but for using it from the shadow Clojure repl.
+
+  Arguments are the same as the Clojure version but `config` needs a `:build-id`"
+
+  [var-symb config]
+
+  (hansel/uninstrument-var-shadow-cljs var-symb config)
+  (dbg-api/publish-event! :cljs
+                          (rt-events/make-var-uninstrumented-event (name var-symb)
+                                                                   (namespace var-symb))
+                          config))
+
+(defn instrument-namespaces-cljs
+
+  "Like `flow-storm.api/instrument-namespaces-clj` but for using it from the shadow Clojure repl.
+
+  Arguments are the same as the Clojure version but `config` also accepts a `:build-id`"
+
+  ([prefixes] (instrument-namespaces-clj prefixes {}))
+  ([prefixes opts] (dbg-api/instrument-namespaces :cljs prefixes opts)))
+
+(defn uninstrument-namespaces-cljs
+
+  "Like `flow-storm.api/uninstrument-namespaces-clj` but for using it from the shadow Clojure repl.
+  Arguments are the same as the Clojure version but `config` also accepts a `:build-id`"
+
+  [prefixes config]
+  (dbg-api/uninstrument-namespaces :cljs prefixes config))
 
 (defn- runi* [{:keys [ns flow-id thread-trace-limit tracing-disabled? env] :as opts} form]
   ;; ~'flowstorm-runi is so it doesn't expand into flow-storm.api/flowstorm-runi which
   ;; doesn't work with fn* in clojurescript
   (let [wrapped-form `(fn* ~'flowstorm-runi ([] ~form))
         ns (or ns (when-let [env-ns (-> env :ns :name)]
-                    (str env-ns)))]
+                    (str env-ns)))
+        hansel-config (tracer/hansel-config opts)
+        {:keys [inst-form init-forms]} (hansel/instrument-form hansel-config wrapped-form)]
     `(let [flow-id# ~(or flow-id (-> form meta :flow-id) 0)
            curr-ns# ~(or ns `(when *ns* (str (ns-name *ns*))) (-> env :ns :name str))]
-       (binding [inst-rt/*runtime-ctx* {:flow-id flow-id#
-                                        :thread-trace-limit ~thread-trace-limit
-                                        :tracing-disabled? ~tracing-disabled?}]
-         (tracer/trace-flow-init-trace flow-id# curr-ns# (quote (runi ~opts ~form)))
+       (binding [*runtime-ctx* {:flow-id flow-id#
+                                :thread-trace-limit ~thread-trace-limit
+                                :tracing-disabled? ~tracing-disabled?}]
+         (tracer/trace-flow-init-trace {:flow-id flow-id# :form-ns curr-ns# :form (quote (runi ~opts ~form))})
 
-         (~(inst-forms/instrument opts wrapped-form))))))
+         ~@init-forms
+
+         (~inst-form)))))
 
 (defmacro runi
 
@@ -199,32 +258,6 @@
   [opts form]
 
   (runi* (assoc opts :env &env) form))
-
-(defn instrument-forms-for-namespaces
-
-  "Instrument all forms, in all namespaces that matches `prefixes`.
-
-  (instrument-forms-for-namespaces prefixes opts)
-
-  `prefixes` is a set of ns prefixes like #{\"cljs.compiler\" \"cljs.analyzer\"}
-
-  `opts` is a map containing :
-       - :excluding-ns  a set of strings with namespaces that should be excluded
-       - :disable a set containing any of #{:expr :binding :anonymous-fn}
-                  useful for disabling unnecesary traces in code that generate too many traces
-       - :verbose? when true show more logging
-  "
-
-  ([prefixes] (instrument-forms-for-namespaces prefixes {}))
-  ([prefixes opts] (dbg-api/instrument-namespaces prefixes opts true)))
-
-(defn uninstrument-forms-for-namespaces
-
-  "Undo instrumentation made by `flow-storm.api/instrument-forms-for-namespaces`"
-
-  [prefixes]
-  (dbg-api/uninstrument-namespaces prefixes true))
-
 
 (defn cli-run
 
@@ -274,7 +307,7 @@
 
       (log (format "Instrumenting namespaces %s" ns-to-instrument))
       (time
-       (instrument-forms-for-namespaces ns-to-instrument inst-opts))
+       (instrument-namespaces-clj ns-to-instrument inst-opts))
       (log "Instrumentation done.")
       (time
        (if flow-id
@@ -283,11 +316,52 @@
       (log "Execution done, processing traces...")
       (log "DONE"))))
 
+(defmacro instrument* [config form]
+  (let [env &env
+        compiler (inst-utils/compiler-from-env env)
+        ;; full-instr-form contains (do (trace-init-form ...) instr-form)
+        {:keys [inst-form init-forms]} (hansel/instrument-form (merge config
+                                                                      (tracer/hansel-config config)
+                                                                      {:env env})
+                                                               form)]
+
+    (if (and (= compiler :clj)
+             (inst-utils/expanded-defn-form? inst-form))
+
+      ;; if we are in clojure and it is a (defn ...) or (def . (fn []))
+      ;; add a watch to its var to track when it is being instrumented/uninstrumented
+      (let [var-symb (second form)]
+        `(do
+
+           ~@init-forms
+           ~inst-form
+
+           (let [v# (var ~var-symb)
+                 [vns# vname#] ((juxt namespace name) (symbol v#))]
+             (rt-events/publish-event! (rt-events/make-var-instrumented-event vname# vns#))
+             (add-watch v#
+                        :flow-storm/var-redef
+                        (fn [a1# a2# fn-before# fn-after#]
+                          (cond
+
+                            (and (:hansel/instrumented? (meta fn-before#))
+                                 (not (:hansel/instrumented? (meta fn-after#))))
+                            (rt-events/publish-event! (rt-events/make-var-uninstrumented-event vname# vns#))
+
+                            (and (not (:hansel/instrumented? (meta fn-before#)))
+                                 (:hansel/instrumented? (meta fn-after#)))
+                            (rt-events/publish-event! (rt-events/make-var-instrumented-event vname# vns#))))))))
+
+
+      `(do
+         ~@init-forms
+         ~inst-form))))
+
 (defn read-trace-tag [form]
-  `(dbg-api/instrument* {} ~form))
+  `(instrument* {} ~form))
 
 (defn read-ctrace-tag [form]
-  `(dbg-api/instrument* {:tracing-disabled? true} ~form))
+  `(instrument* {:tracing-disabled? true} ~form))
 
 (defn- read-rtrace-tag* [config form]
   (let [full-config (merge config