+++ src-inst/flow_storm/instrument/forms.cljs
@@ -1,3 +1,0 @@
-(ns flow-storm.instrument.forms)
-
-;; Just for tests